SNI: using the ClientHello callback.
The change introduces an SNI based virtual server selection during
early ClientHello processing.

Switching to the ClientHello callback sets a reasonable processing
order for server name and session resumption, making it consistent
with BoringSSL behaviour.

A session resumption callback is now invoked after a virtual server
is chosen by SNI, making it possible to resume in a non-default SSL
context.  This is in particular important if verification of client
certificates is enabled.  The change supersedes the fix made in
46b9f5d38 for TLSv1.3 resumed session.

To keep the $ssl_server_name variable working in TLSv1.2 resumed
sessions, as previously fixed in fd97b2a80, a dummy server_name
callback is used.

The ClientHello callback is available since OpenSSL 1.1.1.  For
older OpenSSL versions, a historic behaviour is preserved.